+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Support for loading the data from a 'ng_trajectory' csv file.
 - Callback for `input_file` so that the file can be changed during runtime.
 - Subscriber for `move_base_simple/goal` to rotate and translate the trajectory.
+- Parameter `path_length` to change the number of points in the generated path.
 
 ### Changed
 - Node is now executable using Python 3 as well.